Day 23 part 01: minor bug and too slow
It's nearly there, but it is too slow for the real input (it barely
finishes for the sample, which is far easier to solve).

It's also buggy, because it may exit too soon: as soon as a solution
is found. But perhaps some other edge node will be faster, even
though it is lagging behind in cost.

Needs improvements.